[core] Add more message when accessing a dead actor. (#34697)
## Why are these changes needed?
When any ray worker access the actor which doesn't belong to the cluster, it'll crash. This is really bad user experience because:

- The user will lose ray worker or driver.
- The error message is not very useful.

This PR fixed it by avoiding crash and throw exception instead. When the user accidentially access an actor which doesn't belong to this cluster, it'll get an Exception with message: `It might be dead or it's from a different cluster`.

This PR also deleted some code that's not useful.