[data] [streaming] Fix backpressure when reading directly from input datasource (#34809)
The size_bytes of the input metadata was not properly included from read tasks. This meant we can easily launch too many tasks and go over the memory limit on reads of this type.

This bug only happens on code of the form:

ds.read_datasource().iter_batches().

But it doesn't happen if there are any intermediate transformations, which hits a different code path that has correct metadata propagation.