[core] Start the synchronization connection after receiving all nodes info. (#34645)
There is a race condition between resource broadcasting and node registration. When the node is registered, it'll start the sync connection immediately. If the first message from the GCS is received before the Raylet got all node infos, it's going to drop the message from the node which it doesn't know.

The newly started raylet will only see its local resource which might cause some problem if it's going to schedule some task around.

This PR move the connection to the place after getting all nodes infos.