Deflakey test advanced 9 (#34883)
Previously a bug was fixed in [PR](https://github.com/ray-project/ray/pull/33311) where pubsub causes the leak. Somehow the fix has race conditions and got triggered later when code changes.

The test is flakey because there is a race condition between raylet sending node failure and core worker exit itself.

When disconnect is sent to Raylet, Raylet will start to report worker failure. But the worker still continue to run.

GCS uses worker failure to close the connection. But if the worker is still alive, the worker might send another request the GCS which will lead to the FD leak.

This PR did two improvements:
- Move the heavy workload before sending disconnect request
- Raylet will report worker failure if the socket is closed.