[core] Reduce self alive check from 60s to 5s. (#34992)
This PR reduce the liveness check from 60s to 5s. It also fixed a bug in the old code where if the server is restarted locally, it'll mark the current one as unhealthy incorrectly because the set is not multi set. 

One example,

- If the head node restart in-place
- when it started, it'll mark the old raylet as dead
- in the liveness check endpoint here, it'll mark the raylet as dead because they share the same ip+port.
- then the head node will exit

Node id should be used for this check in the future for simplicity but the correctness is ok, given: No two raylets can start at the same address (ip+port).