[core][dashboard] Make actor tasks'name default to <actor_repr>.<task_name>  (#35371)
We updated the dashboard to show actor's name with actor repr, and we should do the same for actor tasks.

Caveat:

Actor.__init__ is currently not showing up as repr_name.__init__ because during creation task initialization, we haven't initialized the actor states yet, so the repr func should not yet be called. There's workaround (i.e we could modify the init task name later at rendering, but chose not implement in this PR), but would be rather hacky. The major issue is that repr info for an actor is only available on the executor, and on the submitter (or encoded in actor handle). Given the feature is only for dashboard, I feel this the intrusive change in this PR is better.